+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.2.0 (Next)
 - Add functions: truncate, exp, log, next, sqrt, translate.
+- Add fact: powershell_version.
 
 # 1.1.1
 - Expand `script` DRT functionality.